Contact change int to long

+++ features/user/chat_common/src/main/java/com/tokopedia/chat_common/domain/pojo/GetExistingChatPojo.kt
@@ -52,7 +52,7 @@ data class Contact(
         val role: String = "",
         @Expose
         @SerializedName("userId")
-        val userId: Int = 0,
+        val userId: Long = 0L,
         @Expose
         @SerializedName("shopId")
         val shopId: Long = 0L,
